Moved database types into database/types
