+++ app/assets/js/src/data/shared/types/index.ts
@@ -1,10 +1,3 @@
-export type { DatabaseData } from './DatabaseData';
 export type { DataSource } from './DataSource';
 export type { ExportData } from './ExportData';
 export type { ExportDataLike } from './ExportDataLike';
-export { type LegacyExportData, isLegacyExportData } from './LegacyExportData';
-export type { TaggedLegacyExportData } from './TaggedLegacyExportData';
-export {
-	isLegacyExportDataV1_0_0,
-	isLegacyExportDataV2_0_0,
-} from './LegacyExportDataVersions';
