+++ app/assets/js/src/database/LegacyExportDataVersions/LegacyExportDataByVersion.ts
@@ -1,0 +1,8 @@
+import type { TaggedLegacyExportData } from '../types/TaggedLegacyExportData';
+
+/**
+ * Utility function for retrieving a specific version of {@linkcode LegacyExportData}.
+ */
+export type LegacyExportDataByVersion<
+	V extends TaggedLegacyExportData['schemaVersion']
+> = Extract<TaggedLegacyExportData, { schemaVersion: V; }>['data'];
