+++ app/assets/js/src/database/types/index.ts
@@ -1,0 +1,14 @@
+export type { DatabaseData } from './DatabaseData';
+export {
+	isLegacyExportDataV1_0_0,
+	isLegacyExportDataV2_0_0,
+	type LegacyExportDataV1_0_0,
+	type LegacyExportDataV2_0_0,
+	type LegacyStatusName,
+	type LegacyExportDataByVersion,
+} from './LegacyExportDataVersions';
+export {
+	isLegacyExportData,
+	type LegacyExportData,
+} from './LegacyExportData';
+export type { TaggedLegacyExportData } from './TaggedLegacyExportData';
